+++ sdk/php/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 API documentation [Changelog](https://www.kucoin.com/docs-new/change-log)
 
-Current synchronized API documentation version [20250529](https://www.kucoin.com/docs-new/change-log#20250529)
+Current synchronized API documentation version [20260904](https://www.kucoin.com/docs-new/change-log#20260904)
+
+
+## 2026-09-12(PHP 0.1.4-alpha)
+- UTA interface update
 
 ## 2025-06-11(PHP 0.1.3-alpha)
 - Add compatibility for PHP versions from 7.4 to 8.x